requirements.txt
anthropic>=0.40.0
chromadb>=0.5.0
pymupdf>=1.24.0
sentence-transformers>=3.0.0
click>=8.1.0
pyyaml>=6.0
rich>=13.0.0
flask>=3.0.0
python-docx>=1.1.0
markdown>=3.6
bibtexparser>=1.4.0
openpyxl>=3.1.0
gunicorn>=22.0.0
gevent>=24.0.0
rank-bm25>=0.2.2
google-genai>=1.0.0
